fix: studio tests not finding fixtures and weird equality checks
