+++ tests/connectors/studio/conftest.py
@@ -1,10 +1,12 @@
+from collections.abc import Sequence
 from unittest.mock import Mock
 from uuid import uuid4
 
 from dotenv import load_dotenv
+from pydantic import BaseModel
 from pytest import fixture
 
-from intelligence_layer.connectors.studio.studio import StudioClient
+from intelligence_layer.connectors.studio.studio import StudioClient, StudioExample
 
 
 @fixture
@@ -19,3 +21,28 @@ def studio_client() -> StudioClient:
 @fixture
 def mock_studio_client() -> Mock:
     return Mock(spec=StudioClient)
+
+
+class PydanticType(BaseModel):
+    data: int
+
+
+@fixture
+def examples() -> Sequence[StudioExample[PydanticType, PydanticType]]:
+    return [
+        StudioExample[PydanticType, PydanticType](
+            input=PydanticType(data=i), expected_output=PydanticType(data=i)
+        )
+        for i in range(2)
+    ]
+
+
+@fixture
+def many_examples() -> Sequence[StudioExample[PydanticType, PydanticType]]:
+    examples = [
+        StudioExample[PydanticType, PydanticType](
+            input=PydanticType(data=i), expected_output=PydanticType(data=i)
+        )
+        for i in range(201)
+    ]
+    return examples
